runtime: pin Modal image + deps; fix prefill Int64→Int32 loop bounds
Goal: Modal bench numbers and the bare scoring runner must agree. The
previous divergence came from build-surface drift (implicit nvcc
defaults, mismatched package versions). This commit pins every layer
that can move SASS.

config.toml — pin deps (matches flashinfer-ci-cu132:latest @ 2026-04-18):
    nvidia-cutlass-dsl == 4.4.2
    cuda-python        == 13.2.0
    apache-tvm-ffi     == 0.1.10

docs/runtime-pin.md — full runtime inventory captured from the Modal
container, with the base-image digest
  `flashinfer/flashinfer-ci-cu132@sha256:87e3b10d657b4c3f31fd028b01748d82fe09f41a30d820b1afb39ef77fbfa453`
and the reproducibility recipe for bare Docker.  The image digest can
be wired into `MODAL_IMAGE_SHA` in `scripts/run_modal.py` to force
Modal to use the identical cached image — no surprise `:latest` drift.

solution/python/msinfer_entry.py — prefill: cast cu_seqlens[...] to
Int32 before the `for t in range(seq_start, seq_end)` token loop. The
CuTe DSL dynamic-range loop requires Int32 bounds; cu_seqlens is int64,
so `DSLRuntimeError: expected Int32 for start, got Int64` was blocking
the first compile. Matches how flashinfer's own reference CuTe prefill
handles the cast.

Verified on Modal B200:
  Prefill (5 smallest workloads): 0.043-0.059 ms, 34-65× speedup, PASSED
  Decode already validated in the previous commit (30 workloads, ~172× mean)

+++ solution/python/msinfer_entry.py
@@ -222,8 +222,9 @@ def _gdn_prefill_dev(
     qk_head = v_head // (HV // HQ)
     row = tid
 
-    seq_start = cu_seqlens[seq_idx]
-    seq_end = cu_seqlens[seq_idx + 1]
+    # cu_seqlens is int64; the DSL dynamic-range loop requires Int32 bounds.
+    seq_start = cutlass.Int32(cu_seqlens[seq_idx])
+    seq_end = cutlass.Int32(cu_seqlens[seq_idx + 1])
 
     A_log_val = cutlass.Float32(A_log[v_head])
     dt_bias_val = cutlass.Float32(dt_bias[v_head])
